First functional prototype with most major features (except delete/move). Most everything in the Python side should be fairly solid. Most of the stuff in the JS side is likely to change. I will most likely modify the JS to be knockout, because it's messy for what little it does and I need to do a lot of front-end polish.
What is still missing from the back-end is ensuring that there are no views that can access folders aside from the dashboard. I know of some searches (such as the add pointer search in the project) that will still find folders. Also, you can still view a folder in the project view.