Removes redundant date observable extension.
https://github.com/CenterForOpenScience/openscienceframework.org/issues/838

Same as the last one, but for some more fields.

One of the two problems, and the one that made these controls the most unusable, is that the date validator required YYYY-MM-DD, but putting in a date that didn't naturally have double-digits on both MM and DD would strip out the leading 0, thus causing the validator to fail.

There were two extensions to the observable that were date related. One was going to the validator, and the other… I'm not sure what it was doing. Taking it out didn't seem to break anything, and fixed at least the most important part of the issue, so we'll start with that.