Adds back the extra observeable and makes the date printer not break
https://github.com/CenterForOpenScience/openscienceframework.org/issues/838

The date printer was not padding the number with 0s, which apparently works fine for JS Date() on Chrome, but not Safari and firefox. I added a padding function and padded the month and date as necessary.

The current version will not reformat the date when saving, so you could put in July 5, 14 and 07-24-2014 as your start and end dates and they'll be shown like that until the page is reloaded. It all works properly, though, and checks for the end date being before the start date or putting in something that's not really a number.