Added back in asDate to jobs (not just education). Changed default delim to /
https://github.com/CenterForOpenScience/openscienceframework.org/issues/838

When formatting the dates, the previous version was converting everything to YYYY-MM-DD, which was mostly cool and fairly readable, but the one failing case right now is using a combination of single digits for months and days with dashes as delimiters.

So now, the formatting converts it to YYYY/MM/DD. Overall, it's a little finicky if people use dashes, but everything we do encourages them to do things in a way that will work, and it won't take much experimentation to put in a date. Ultimately, I think we'll need to replace freeform text with either fields or a date picker of some sort.