Modified Folder Creation to work like new Project Creation Refactoring
https://github.com/CenterForOpenScience/osf/commit/53b4875ac80e22e0aaa2db387e65f82e6148f9f3#diff-033d4b3df3102b7d1aace074938585e9R119

We stopped using WTForms for creating projects, and I was creating folders (at least, in one instance) in the same way. We'll probably change this again once Alex's onboarding stuff is up.

There are a couple of changes to the Project Creation, mostly on how I'm handling an empty folder name. Instead of disabling the submit control, I'm popping up a message saying that we need a name.