improve front-end formatting
this commit contains several changes to the css and
mako files which improve the general look of the search page.

Primary among these changes is a fix that prevents text form flooding
into other fields on a thin screen, the use of a common css class
for almost all fields, and the addition of a no_ statement for
user job/school fields.