improve "no query" behavior
This commit improves the logic for what to show at the top of the
search page. This should make it much easier for future editors
to make changes to this part of the page. It also makes the page's
behavior more uniform, which I will use to add features that apply
in all cases in the future.