fix elasticsearch tests
There were some tests failing in test_elastic.py because
they were using a function that was set for the old way
of searching, without using dictionaries. Thus, I replaced
that with a working version and also made a couple of other
changes to match with the new way that search works, for
example moving a quotation mark that no longer works.