fix unrelated tests
For some reason beyone my knowledge, a couple of tests,
as well as a docstring in core.py, were acting as if
'position' should be used as a key in a user's job instead
of 'title'. This was causing tests to fail unnecessarily,
so I replaced the 'position's with 'title's.

note: the reason that the tests didn't fail before is that
nothing was ever trying to call the 'title' key upon saving
before I added it to elasticsearch, so the test didn't
have any reason to think that 'position' was the wrong key.