Fixed by using the same back-end gravatar calls as elsewhere in the system and removing on error view code.
https://github.com/CenterForOpenScience/openscienceframework.org/issues/842

This was fixed in a feature branch https://github.com/CenterForOpenScience/osf/pull/834 but was fixed on the view rather than the serialization in python. I've moved the fix back a bit farther after talking with @sloria. This should work much more consistently and make for easier maintenance.

Currently, this is not showing the gravatar at the requested 120 pixels; I presume that's a CSS thing or similar.