Make tag searches have to be exact
Previously, Elasticsearch was analyzing tags just like
any other field. This meant both that random similar
tags might appear in a tag search, which should be exact,
and also that tag searches with whitespace wouldn't work
properly.
However, the current way that we create the Elasticsearch index
doesn't allow us to create customized mappings, so I added
a section to the migrate_search command as well as the other
parts of search that will initialize the index so that we can
customize it as much as we want before it defaults using the
node and user docs.