Fix some failing tests
There were some tests that were failing due to the way that
tags worked, and while my earlier change fixed the problem in
one way, it also caused a different issue in that the tests
did not yet initialize the index.
This commit changes that, by having all tests that involve search
completely delete and reinitialize the search index.