Make word order in search not matter
Previously, due the the fact that searches used a phrase-prefix
filter, the order of the words in a query mattered. This meant
that searching the words in the wrong order would result in a
failed search. I fixed this by removing the phrase-prefix
filter, which will make searches take slightly longer, but fixes
this bug.