Make tags under projects add to query
Previously, the tags that appeared under projects, components
and registrations would create a new query, rather than adding
to the existing one. This made sense when the tag cloud showed
all tags from the page, since there was already a way to add
to the query, but now, it makes more sense to simply have both
act the same, especially since users can just click the "x"
next to their original query to remove it.