Show expected warning when uploaded file is too large.
The "uploadError" callback currently expects the "message" parameter to
be a JSON-encoded string, since OSF errors are passed back as JSON. But
errors from Dropzone are plain strings. This commit first attempts to
parse errors strings as JSON, then degrades to displaying the raw
string, to handle both OSF errors and Dropzone errors.

[Resolves https://github.com/CenterForOpenScience/openscienceframework.org/issues/870]