Handle not having a user on the last log in both back and front end.
The project organizer currently grabs the last log item to find out who made the last change to a project, then serializes the last name. Some log items do not have users to attribute, because they were done by system tasks. In these instances, I pass back an empty string, and the front end does not display the comma or the name, as there is nothing to display. It will only show the last modified date.