Shrink Project Organizer from "100%" of the div to 98% to ensure it doesn't overflow.
This fixes an interaction problem between input devices that don't have horizontal scrolling and Chrome. When set to 100%, the hgrid would appear as a 651pixel div inside of a 650-pixel div. Because the logitech mice we use don't have horizontal scrolling, when one is plugged in, Chrome would foist a horizontal scrollbar onto the slightly-too-large hgrid.