Disable the contextual menu on the page that holds the Project Organizer.
This is a fix for https://github.com/CenterForOpenScience/openscienceframework.org/issues/945

The side effect is that you can't have a contextual menu on the Dashboard page, but it doesn't otherwise break anything that I've seen. You can still bring up dev tools and things, but you have to do it from normal menus and not contextual menus.