Fix missing subprojects in smart folders.
The previous version of the smart folders search views were not grabbing sub projects, because the component search was overly restrictive. Instead of not grabbing folders, it avoided grabbing projects. Once that was in, I also had to make sure that it wasn't trying to grab anything that was a top-level project. This seems to catch all of the important cases.

https://github.com/CenterForOpenScience/openscienceframework.org/issues/957