Make exclusion of search users happen in query rather than after.
Previously, we were going through the search results in python and removing the users that had contributed. Now, we are adding those users to the search query as part of a 'must not' filter, so they will not be considered in the search. This will keep the number of users displayed in the project up to the maximum number displayable if they are available, rather than the maximum number displayable minus those that were already in the project.