Send the proper mime type with standard file types so Safari will download properly.
Still does not properly address the main part of https://github.com/CenterForOpenScience/openscienceframework.org/issues/382 which is markdown files, but fixes a lot of cases that we didn't realize were there (i.e. everything but markdown). And, indeed, markdown works better because it'll be downloaded as text instead of html. But still.