Add python-magic as a second layer of mimetype checking.
This moves the mimetype checking into its own util file and adds the ability to use python-magic to examine the content stream to try to find the type. If you don't have libmagic installed, it will fail gracefully and do no worse than the default case before.

Still to do is to add in the custom mime mapping file.