Add custom mime mapping file with markdown extensions.
At this point, if we ever need additional mime types added, we just have to add them to the mime.types file. Also, we should be able to find someone who is maintaining such a file with all the new file formats so we can just add them to the mime.types file as needed, and/or one day find a project that does all that for us that isn't so manual.