Use parse_date instead of strptime to avoid rare error.
When isoformat is called on a datetime object with no floating point
value, the resulting string has no floating point value. This breaks
stprtime when the format string includes a float point formatter. The
result is that tests will fail on parsing in the very rare event that
the test object is created when the floating point component of the time
is zero. This patch uses the more robust parser in dateutil to avoid
this edge case.