Raise 401 on private views when not logged in.
When a user tries to view a private page while not logged in, the OSF
currently returns a redirect to the login page, which resolves with a
200 status code. This is semantically incorrect; the user should see a
401 instead. This behavior also results in a number of incorrect tests
passing: if tests do not pass the correct authorization to the
application, a 200 code may still be received. This patch causes the
login page to throw a 401 when reached as a redirect from a private page
when the user is not logged in. Incorrect tests that were passing
previously (but should have been failing) have also been fixed.