Handle revoking dropbox token from db website
When auth token is revoked from dropbox account
settings, the OSF project and account settings page
do not display errors and it appears as if the
folder is still linked.

Display message to user when settings page is
loaded indicating that dropbox cannot be reached
and that they may need to reauthorize.