Monkey patch dropzone to not upload directories.
This only works well in Chrome, which is the only browser that supports dropzone's directory uploads, because Chrome is the only thing that detects directories properly. However, it should under all circumstances keep directories from uploading, which will fixes https://github.com/CenterForOpenScience/openscienceframework.org/issues/756 and fixes https://github.com/CenterForOpenScience/openscienceframework.org/issues/754 (as well as a problem with the onboarding.

That being said, HGrid swallows the errors from the directory uploads because they happen so quickly. There may need to be an additional pause to make it more obvious that this is happening.

With Firefox and Safari (and likely IE), the error is that the server returned a 0. Front-end changes might check for that error and report something different to make this more user friendly in those circumstances. As things stand, though, this shouldn't introduce any new errors, and should work better or the same as existing configurations.