Store destination data on file objects.
Dropzone uses top-level variables for upload method and URLs, and HGrid
sets both variables on drop. An unintended consequence of this behavior
is that dropping a file while previous uploads are pending changes the
destinations of the pending files. This patch stores destination
information on each added file object and patches Dropzone to pull
information from files if present. Note: This solution is less than
ideal and should be replaced with a pull request to Dropzone.

Resolves https://github.com/CenterForOpenScience/openscienceframework.org/issues/717