Do not show addon-success message with invalid credentials
When a user successfully authorizes dropbox on their
account settings page, a success message is shown
with "next steps" until a user configures the addon
for a project. If credentials are revoked before then,
an error message is shown in addition to the
success message. Change to only show success message
if credentials are valid.