Check if valid credentials in serialize_settings and show error msg
Currently, if figshare access is revoked externally (not through
the OSF) there is no indication that this occured except that
figshare files do not show up in HGrid.

Call get_options(), which makes a request to
http://api.figshare.com/v1/my_data/projects and /articles and
returns '401' if user is unauthorized. If credentials are
invalid, display error message to the user.