Move addon success message to dropbox_user_settings.mako
The success message shown when a user successfully
configures an addon stays there until a project/node
is configured. However, if dropbox access is revoked
the success message will still be displayed since the
logic is in a separate location (addon_permissions.mako)

To fix, handle dropbox user settings messages in
dropboxUserConfig.js.