Replaced quoted text that would break the GrowlBox with text that would work.
GrowlBox, when used as a push_status_message replacement, can only take text that is in one line, because it is building the JS for it dynamically in a mako file, and the formatting will cause improper end of line if there's a break.  If the line breaks are on the outside of the string, then no problem; trim will take care of it. If the line breaks are within a string, then it's a problem, so I replaced all the multi-line strings that get sent through push_status_message into either single-lines or tuples of strings that get auto-concatenated.