Show correct dates in file revision widget.
We currently show the date created metadata from Rackspace, which will
be different from the date the user uploaded the file when the same file
is uploaded to different nodes at different dates. This patch shows the
date at which the user uploaded the file instead. Thanks @caileyfitz for
reporting.