Test that 'subscribe' and 'profile' webhook actions do not update user
If mailchimp webhooks are set up to notify on changes made
via the API, it will send 'subscribe' and 'profile' actions
to the OSF as well as 'unsubscribe.' Previously, the 'profile'
action unsubscribed users in the OSF db.

Add regression tests to ensure that 'subscribe' and
'profile' actions do not modify the OSF db.