Force a user to go to user settings page in order to update a profile
Users who change their name on their user profile page do not see that there is an additional step of updating the name parts, and this has caused confusion on multiple occasions such as when Lyndsy was demoing the citation modifications and didn't realize that the name needed changing in multiple ways. This eliminates the ability for a user to make profile modifications on the profile page, and instead takes them to the settings page for such changes.  I had to make some changes to the knockout which wasn't really expecting to have a situation where you would have a view-only version of the page even if the user had the capability to edit their own info. It was almost there, but there were some assumptions based on what we had currently implemented that were not correct.