Only notify indirect subscribers on parent who can access component
When an event happens on a component, the parent project
subscription is also checked in 'check_parent' for
any indirect subscribers (e.g. user subscribed to
parent project and component remains the default
'adopt parent')

Add check to ensure that emails are only sent
to subscribers on the parent project if they also
have access to the node where the event occurred.