Keep displayText from updating if view panel is not visible
This fixes a bug with a new interaction added recently showing compare to live preview if edit is on and view is turned off. There shouldn't be any way to change the version of the view panel outside of that mode if view is not visible. Also simplifies visibility logic greatly.

[skip ci]