Include parent title for nodes in email transactional template
If an event happens on a node, the node title is given but
there is no context as to which project it belongs to. If
the component has a commonly repeated name such as "Data"
or "Results" the notification is not particularly informative.

Add the parent title to the template if the notification
node is a component.