Update migration to handle dates for pre-migration files.
After running this script last time, we changed which fields on the
`OsfStorageFileVersion` model were displayed to the user:
`date_created` now reflects the time the file was uploaded (which is
displayed to the user), and `date_modified` reflects the time the file
reached the backend (which does not always correspond to the upload
time, and is not displayed to the user). The original migration to OSF
Storage didn't copy creation time correctly; this updated migration
copies times from the `nodefile` collection to the `date_created` field
of `OsfStorageFileVersion`.

[Resolves #1885]