Fix project settings page anchor scroll issues properly without needing jQuery
The anchor scroll in settings page was broken because 0.24 extended the fixed navigation at the top so the scroll would be hidden underneath the fixed section. There was already a fix for this but I didn't realize that at the time and used a jQuery solution which is not necessary. This commit fixes that problem by adjusting the original fix and removing jQuery.