Random color generator and new color generator option
Generates a color based on ints first, then does the distance calculation and then converts to hex. Also add an optional function that we may tweak later that converts the name of each source to a color so that we have persistance of colors even when we add a new provider.