Fix message shown when no registrations on component
Fixes issue where the message "To register this
component you must register the parent project"
was shown to the public and to contributors who
do not have permission to register the parent
project.

Only show the message to users who are admin
on the parent project, since they are able
to register it. For everyone else, show the message,
"There have been no registrations of the parent
project (<parent project title>)."

Closes #1628