Fix #390
Use knockout and selects to make it impossible to move the contributor
bar when select is clicked on.

An error was fixed by unknown updates in the meantime. Also restrict
what is passed into python to mirror the tests of contributors to
four variables.

Amend: remove console.logs, get around amend.