Fix osfstorage and box file highlighting
OSF storage files now use an id for
the path instead of the full path,
breaking the file highlighting.

Use window.contextVars.file.extras.fullPath
instead of window.contextVars.file.path to
get the proper named path. Since this was
added for box files, specialcase that as
well so that box files highlight on the
file detail page.