Basic basic waterbutler integration
I still need to work out the basic structure of the nodeFilesSerializer serialization, as well as decide if I'm just going to proxy all the WB stuff short of the file download or convince them to HATEOAS in the same way we do. But I am creating a WB link that works properly to get the basic provider information.