Fix bug in oop.js
When multiple classes are defined in the same file using defclass, the
noop acted as a global variable in oop.js, resulting in the constructor
being shared amongst all classes in the file.

This was exposed by defining two classes that shared a `submit` method.
Even when the second class was not exported from the module, the first
class's `submit` method was overwritten by it.