Make API grab metadata from Waterbutler and provide Waterbutler routes for CRUD operations
One of the keys to the current implementation plan is going all in on HATEOAS; this is consolidating a couple of APIs into one while allowing an external reference to an API. What this means is that the pattern for usage of this api will lean against building URLs by hand and towards following and getting links from our link objects. This will allow Waterbutler to change without breaking the main API.

Along with doing this, I had to rework _get_attr_from_tpl a little. It worked great on the main object representations, but not as good at anything that wasn't a list of Nodes or Users. I am now having it check to see if the attribute is directly in the object rather than just failing out immediately.