Retraction/Embargoes now consider each other
Retractions will now cancel any pending or active embargoes
of registrations when they are approved.

Embargoes of registrations that did not exist before the embargo was
initiated delete that registration if the embargo is disapproved. If
the registration already existed (e.g. registrations created prior to
release of the embargo feature), the registration not be deleted.