Update pending_embargo.txt.mako
Made the same very minor edit as done for retractions.  The same change there would be good here - replacing "another administrator" with the name of the administrator.

Also, the same text could be used with "embargoed" removed for an immediate registration.  OR, this text could be edited so that it works for both embargoed and non-embargoed registrations.  I think that would just be removing "embargoed" from the opening sentence and then having the embargo end date be 48 hours after registration for "immediate" releases.