Treat get_comparison_operator more like a dictionary
This is not as general purpose, but Steve thinks it's likely easier to read and use than the other, and the other would be handling some rare use cases, so we'll go with this.

Also, there was a bug in the handling of boolean field types so that if you had a project with a title that you wanted to seach for 1, 0, true, or false on, it would convert those values to true or false even if the field type wasn't boolean. So that's fixed.