Modularize it
Move repeating code into a mako module and a ko module. The
reason for using mako is because it builds the elements into the
served webpage, however with the visible stuff for the select list
it needed to be moved into a ko template so as to not load then
disappear, making the loading process ugly.

Add sliding and timing to animations.

Tweak a few styles to apply correctly.

[skip ci]