Embargo disapprovals now return proper redirects
Previously, upon following an embargo disapproval link they were
forwarded to the registrations page. If this was a temporary registration
e.g. one created pending the approval of an embargo, it would direct users
to a resource deleted page.

Now the redirect will only return the user to the registration an embargo
was attached to if it was pre-existing and therefore not deleted upon
disapproval of the embargo. Otherwise, it will redirect users toward
the parent node of the registration.