Initial refactor
Refactor file notifications to handle move and copy. Use a pythonic "switch"
statement to address the 6 different options. This isn't the most efficient
number of lines of code as there is a bit of repetition, but it is a trade off.

Looking into cutting some of the repetition. Main focus will be refactoring
subscriptions to work with move and copy, and track down how to do guids
without knowing the provider, or getting the provider some other way.