Fix regression preventing subfolder expansion in citationgrids.
Regression was introduced in 7d8e492c4838daca0e15ef082678ea62c407607f.
The assumption was made that all items in the returned dataset would
have an item[data][csl] value - but folders do not. To resolve, I've
added a conditional that first checks for the existence of the key, and
falls back to the previous method of action.