Move contrib_added signal to node.add_contributors
Node.add_contributor is used in a variety of places
where a user is added before the node is saved for
the first time (e.g. forking, templating, testd) - this
causes errors in building urls for the email
because the node._id is None.

Move signal to add_contributors, which is only used
when adding users to an existing project.