Add addon log templates for deauthorizing with no user
If a user deletes an addon by going to their user settings and
unchecking the addon, the resulting logs on projects will not
have a user associated with them. Therefore, there need to be
templates for deauthorization logs without a user.
In the future, the code should be refactored to attach to the
correct user to the deauthorization logs