Check parent nodes for write permissions
Users attempting to register projects with components might not have
`write` permissions for all components. This will result in a 403 for
all `copyto` actions as well as `copyfrom` actions if the component
in question is not public. To get around this, we have to recursively
check the node's parent node to determine if they have `write`
permissions up the stack.