Discussion overview page should be 'node'
The discussions page endpoint gets the
page (overview, wiki, file) off of the
url args. But in the database, the
overview page is 'node'. Add a check
to make sure the page == 'node' when
the url args say 'overview' otherwise
searching for the comments for the
overview page will turn up nothing.