Do not raise error when throttling emails
If a user is removed and then added to a
project before the throttling period
expired, an exception gets raised and
the user cannot be added to the project.

Replace the raised error with a return
statement so that the user can still get
added to the project and they do not
get any emails during the throttling period.