Implement Organizer Collections endpoints
Organizer Collections are what the project organizer uses to organize projects. The structure of Organizer Collections is much simpler than the v1 implementation and does not support back-end smart folders nor nesting of Collections. We call them Organizer Collections to disambiguate from API collections. All Organizer Collections do is point to nodes. By and large, Organizer Collections are Nodes with most of the functionality stripped out.

NOTE: Everything is set to be in DEV_MODE because we have an internal projec that needs this immediately. Tests are not in yet, but will be put in hopefully while someone is reviewing basic implementation.