Update unread comments in node metadata
Update the node comments metadata to
include the total number of unread comments
on a project, the total number of unread
file comments and the number of unread
comments on the project overview page.

For file comments, make the waterbutler
request in the api context to avoid errors.
As a result this function is both on the
comment model and repeated in the api.