Make contributors anonymous
Note: this is necessary because contributors are sketchy on the node model. There are foreignfields on the node which map directly to contributors, and there are other lists on the model that give the permissions. Rather than, say, giving each contributor its own id, it's just a list of users that eventually were given permissions. Ideally, contributors would be re-worked to not leak user id info, but that's a whole big process on its own, so for now, we'll make the contributor serializer anonymous.

Oh, and right now that doesn't work with embedded contributors (or, I expect, users) because embeds don't include the request context which is necessary in order to get the query parameters which will let me know if there's a private key and if it's anonymous. I'll deal with that tomorrow.