Increase efficiency of code to remove problem entries
I talked with Sara, and the crux of the problem with the script is that there are a bunch of users who have only one log entry, and most/all of those are these people who have only logged in and created a dashboard. To handle that case, I verify that there's at least one log, then check to see if that log is a `dashboard created` log. If it is, I decrement it.

The downside of this approach is that there are some people who will be counted incorrectly as depth users. Those would be easy to figure out. Also, there will be a little variation between the number of log entries in some cases. However, nobody cares about any of these cases. This should take a little longer to run than the version currently in use, but it shouldn't be a significant problem. If need be, we could either make this more accurate or a little faster, but this is probably a fine set of compromises for our current needs.