Add author info when creating comment
When a POST request is made to create a
new comment or reply, the commenter's info
is not embedded in the response.

Get the current user's author info in
CommentListModel (where the comments list
gets fetched) so that whenever the user
makes a comment, their info can be added
to the CommentModel.