Remove file from node.commented_file on delete
When an addon is re-connected, files that were
previously commented on do not exist because that
folder is not connected (but the addon is). As a
result, the files get removed from node.commented_
files which throws off the unread count.

Remove the file_id from node.commented_file when
the file gets deleted. If a file does not "exist"
when checking the unread count, skip it.