Update view to reflect new server logs
The view now updats accordingly from the server.  It was a little tricky
to figure out what parameters needed to be passed to NodeLogs.  I feel
like this was an incredibly hacky solution, basically just push in the
things I needed but it was the only way to really pass things around
correctly.  Anyway Code Complete.

[#OSF-5056]