Implements Preregistration logging
Tricky.  There are some errors at this point that may be expected but as
of yet that's not proven.  Implements the logging behavior the same way
I did it last even if that's going to end up not being best practice.
THE BIG TODO here is implementing a way to have the time display the way
the left hand column works but that will still take some time and maybe
some monkeying around with the JS.

[#OSF-5480]