Fix reporting abuse error
When unreporting a comment and immediately
reporting it again, it fails because the
message is null. Also, the form appears after
unreporting a comment when it should not.

Make self.abuseText default to an empty string
in the request payload. Also, set self.reporting
to false after successfully reporting a comment
so that the form does not re-appear when clicking
unreport.