Fix bug when approving registrations of projects with private components
If the initiator of a registration was not an admin on any of the
project's components, our registration approval script would fail due
to a permissions error.

With this patch, admins on components will be sent an email to
approve or disapprove registrations of the parent project.

paired with @samchrisinger

See OSF-5547