Refactors user merging verification to require pin
This is by no means complete.  This implements the front end of the
application.  It produces a pin and attaches it to that email to send to
someone.  The pin is produced by creating a 6 digit random number.  Also
in this is required a mechanism to verify that the provided pin is
correct; however, this functionality (server and client side) has not
been implemented by this commit.