Update file move notification emails to show file name
When a file move via the file browser takes too long, we let the user
get back to browsing and then send them a notification when it's done.
The notification email was showing the `path` of the moved file, which
is an identifier, not a human-readable filename.  This PR updates
`create_waterbutler_log` to pass the `materialized` property of the
source and destination log payloads instead of the `path` property.

Fixes: [#OSF-5648]