Update language: statistics to analytics
The "Statistics" tab has been renamed to "Analytics." This is updating the language on the Getting Started page.

I am not updating the anchor on the page because I'm concerned links point here and I don't want them to break. The entire Support/Getting Started page will be changed soon and I think we should deal with that issue then.