Update Guid.referent on file move/rename
Currently, whenever the path or node of a file changes,
a new StoredFileNode will be created and the file will
receive a new Guid.

Update the file's  Guid.referent to the new StoredFileNode
so that the Guid of the file remains the same. If the node
of the file changes, update the `node` field for all of it's
comments to the destination_node.