Get guid_list post file/folder move/rename
When a box folder is moved/renamed, it is
not possible to find its contents with just
the folder path. But, since the materialized
path is not available until after wb gets auth
from the OSF, get the list of file guids in
update_file_guid_referent & update the get_guids
method to take in the materialized path.