Move deleted addon files to TrashedFileNode
When an addon file is deleted, make a TrashedFileNode
record and delete the StoredFileNode record. If an
addon folder is deleted, do this for all of its files.

If a filenode that does not exist is being viewed,
move it to TrashedFileNode & remove its StoredFileNode
as well -- this handles files deleted through their
external provider.