fix incorrect file update api docs
The docs claimed that a file could be updated by issuing a PUT request
against the parent directory's endpoint with the query parameter `name`
set to the name of the file.  This is incorrect; files should be updated
by issuing a PUT against the file's endpoint.  The incorrect method may
have accidentally worked in the past, but this will be fixed in an
upcoming release of WaterButler.