take out multidashes from BAPI docs to appease swagger
swagger thinks three or more dashes on their own line indicates the
beginning of a YAML document.  It then dies when the content after
isn't valid YAML.  Replace the dash strings we were using as ascii art
with equal strings to appease swagger.

Also add a note that the BAPI docs are the current preferred doc source.