Remove Node.commented_files
Remove the commented_files field from
the Node class. It is currently used
when calculating the total number of
unread file comments on a project, but
that number is not being used anywhere.
Updating the commented_files field will
also lock the node and the calculation
for total unread file comments is
expensive.