Ignore more celerybeat-schedule files
Since celerybeat uses anydbm, it might have one of a number of file
extensions. It's actually pretty hard to figure out what those file
extensions might be, but it's easy enough to keep them out of git with a
wildcard.

P.S. I don't know what the celery_beat.py.db file is about, so I'm not
touching that.