Remove unused backref from MailRecord model
This model is only used in email submissions to conferences (OSF for meetings), which I can not test locally. Call out to QA to hit this endpoint, but no code references to a use of this backref were found.

For the record: Backref would be pointing at records of type `user` or `node` if present.

[#OSF-5910] [ci skip]