Add missing '/' to start of materialized path
File materialized paths should begin with '/'.
Fix issue where the materialized paths returned
from waterbutler when renaming a file in dropbox,
github, googledrive and s3 did not start with
a '/' and caused errors when renaming the file again.
Because the materialized path did not match the one
used to look up the file, the files received new guids.