Only show comments nav link for pages w/comments
When the page width narrows and the project nav bar
collapses, the comment icon does not show and is
replaced by a "Comments" link in the dropdown. Fix
issue where the "Comments" link was not appearing
on wiki pages. Also ensure that the link is only
available on pages bound to the comment view model
instead of hard-coding which pages should be excluded.