Switch to having default behavior in OsfTestCase
Create a new TestCase that allows for manual creation of Notifications
for more precise testing abilities - put the new default behavior of
creating subscriptions manually back into the main tests because that's
what happens on the real model. This does not seem to significantly slow
down total test suite speed.